requirements.txt
Brotli==1.0.9
certifi==2019.6.16
Click==7.0
dash==1.16.3
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
Django==2.0.3
Flask==1.1.1
Flask-Compress==1.7.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.6.1
MarkupSafe==1.1.1
numpy==1.19.2
pandas==1.1.3
pipenv==2018.11.26
plotly==4.12.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.2
retrying==1.3.3
six==1.15.0
virtualenv==16.7.4
virtualenv-clone==0.5.3
Werkzeug==0.16.0
